Refactor to obtain google api key from env
This is where the RIO pattern shines, imo: all the way down in Geo.hs,
I only care that `env` `HasGoogleApiKey`, and as long as I `runRIO` at
the call site (API.hs), I know that I'm in the monad and can `ask` for
the api key. No need to thread it down to all functions, just keep
them in the same context.